#1325 feat: logic moved to DeleteEntityAction and RemoveAssociatedAction

+++ plugins/BEdita/Core/src/Model/Action/DeleteEntityAction.php
@@ -13,6 +13,7 @@
 
 namespace BEdita\Core\Model\Action;
 
+use Cake\Network\Exception\ForbiddenException;
 use Cake\Network\Exception\InternalErrorException;
 
 /**
@@ -43,6 +44,11 @@ class DeleteEntityAction extends BaseAction
      */
     public function execute(array $data = [])
     {
+        // If object is not deletable returns 403 FORBIDDEN
+        if ((int)$data['entity']['id'] === 1 && ($data['entity'] instanceof \BEdita\Core\Model\Entity\Role || $data['entity'] instanceof \BEdita\Core\Model\Entity\User)) {
+            throw new ForbiddenException(__d('bedita', 'Could not delete "{0}" 1', $data['entity']));
+        }
+
         return $this->Table->deleteOrFail($data['entity']);
     }
 }
